requirements.txt
asgiref==3.6.0
Django==4.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==9.5.0
psycopg2==2.9.7
sqlparse==0.4.4
tzdata==2023.3
